requirements.txt
torch>=1.10.0
torchvision>=0.11.0
Pillow>=8.3.2
pyyaml>=5.4.1
